Moved ExcThread to exception module
